# apps/backend/openapi.test3.yaml
openapi: 3.0.3
info:
  title: X-Ear CRM API (auto-generated)
  version: 0.0.0-auto
servers:
- url: http://localhost:5003
paths:
...
  /api/product-suppliers/{ps_id}:
    put:
      summary: Update a product-supplier relationship
      operationId: suppliers_update_product_supplier
      tags:
      - Product-suppliers
      parameters:
      - name: ps_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
    delete:
      summary: Delete a product-supplier relationship
      operationId: suppliers_delete_product_supplier
      tags:
      - Product-suppliers
      parameters:
      - name: ps_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Success
  /api/products/{product_id}/suppliers:
    get:
      summary: Get all suppliers for a specific product
      operationId: suppliers_get_product_suppliers
      tags:
      - Products
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
    post:
      summary: Add a supplier to a product
      operationId: suppliers_add_product_supplier
      tags:
      - Products
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
...
  /api/suppliers:
    get:
      summary: Get all suppliers with optional filtering and pagination
      operationId: suppliers_get_suppliers
      tags:
      - Suppliers
      responses:
        '200':
          description: Success
    post:
      summary: Create a new supplier
      operationId: suppliers_create_supplier
      tags:
      - Suppliers
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
  /api/suppliers/{supplier_id}:
    get:
      summary: Get a single supplier by ID
      operationId: suppliers_get_supplier
      tags:
      - Suppliers
      parameters:
      - name: supplier_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
    put:
      summary: Update an existing supplier
      operationId: suppliers_update_supplier
      tags:
      - Suppliers
      parameters:
      - name: supplier_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
    delete:
      summary: Delete a supplier (soft delete by setting is_active=False)
      operationId: suppliers_delete_supplier
      tags:
      - Suppliers
      parameters:
      - name: supplier_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Success
  /api/suppliers/{supplier_id}/products:
    get:
      summary: Get all products for a specific supplier
      operationId: suppliers_get_supplier_products
      tags:
      - Suppliers
      parameters:
      - name: supplier_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
...